+++ r3el/constants/DLossPlot.py
@@ -1,7 +1,0 @@
-from typing import Final
-
-
-class DLossPlot:
-    WIDTH: Final[int] = 1500
-    HEIGHT: Final[int] = 450
-    SCALE: Final[int] = 1